## global observer/controller cleanup ##
- merged the o/c changes from master branch
  - adjusted the configuration file for better access to contents
  - adjusted the configuration file to include the modules for the new modular controller
  - added configuration parameters for an enforced scheduling delay at the start of the simulation and between each scheduling
  - updated the default configuration files after these changes
- removed the old global controller

+++ osh_core/datatypes_src/osh/datatypes/power/LoadProfile.java
@@ -254,7 +254,7 @@ public abstract class LoadProfile<C extends Enum<C>> implements ILoadProfile<C> 
         while (nextEntry != null) {
             long nextChange = nextEntry.getLongKey();
 
-            avg += ((double) currentEntry.getIntValue()) * ((double) (nextChange - currentTime) / (double) (end - start));
+            avg += currentEntry.getIntValue() * ((double) (nextChange - currentTime) / (end - start));
             currentTime = nextChange;
             currentEntry = nextEntry;
